Refactor signIn -> logIn, split out register page
